+++ app/src/main/java/de/passwordvault/view/activities/HelpActivity.java
@@ -2,6 +2,8 @@ package de.passwordvault.view.activities;
 
 import android.os.Bundle;
 import android.webkit.WebView;
+import android.webkit.WebViewClient;
+import android.widget.TextView;
 import androidx.annotation.Nullable;
 import java.util.Locale;
 import de.passwordvault.R;
@@ -11,16 +13,35 @@ import de.passwordvault.view.utils.components.PasswordVaultBaseActivity;
 
 /**
  * Class implements an activity that can display a localized help page within a {@link WebView}.
- *
  * The filename of the asset that shall be displayed needs to be passed as extra with key
  * {@link #KEY_PAGE}.
- *
  * For further information about localized assets, see {@link LocalizedAssetManager}.
  *
  * @author  Christian-2003
- * @version 3.5.3
+ * @version 3.5.4
  */
 public class HelpActivity extends PasswordVaultBaseActivity {
+
+    /**
+     * Class implements a custom web view client for the help activity.
+     */
+    private class CustomWebViewClient extends WebViewClient {
+
+        /**
+         * Method is called whenever the web page finishes loading.
+         *
+         * @param view  WebView that finished loading.
+         * @param url   URL that was loaded.
+         */
+        @Override
+        public void onPageFinished(WebView view, String url) {
+            super.onPageFinished(view, url);
+            TextView activityTitleTextView = HelpActivity.this.findViewById(R.id.title);
+            activityTitleTextView.setText(view.getTitle());
+        }
+
+    }
+
 
     /**
      * Attribute stores the key with which to pass the filename of the page to open.
@@ -52,10 +73,12 @@ public class HelpActivity extends PasswordVaultBaseActivity {
             return;
         }
 
+        //Load help page:
         try {
             LocalizedAssetManager assetManager = new LocalizedAssetManager("help", Locale.getDefault());
 
             WebView webView = findViewById(R.id.web_view);
+            webView.setWebViewClient(new CustomWebViewClient());
             webView.loadUrl(assetManager.getFileUri(filename));
         }
         catch (Exception e) {